remove mio_driver package and restore registration.rs
